% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestDeconvolution.R
\name{bestDeconvolution}
\alias{bestDeconvolution}
\title{Find cell-type gene signatures and do deconvolution}
\usage{
bestDeconvolution(
  yref,
  n.types = NULL,
  scree = "drop",
  logTransform = F,
  known.props = NULL,
  methods = c("TOAST", "linseed", "celldistinguisher")
)
}
\arguments{
\item{yref}{matrix, numeric expression matrix}

\item{n.types}{integer, number of cell-types}

\item{scree}{character, method to estimate n.types if n.types is NULL}

\item{logTransform}{logical, T/F if yref is in log-scale}

\item{known.props}{matrix, known proportion matrix, NULL if not known}

\item{methods}{vector, character vector of deconv methods to use}
}
\value{
list with cell-type proportions and expression profiles
}
\description{
The function finds cell-type gene signatures and does deconvolution
using the TOAST framework with NMF
}
