% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigstatsenet.R
\name{bigstatsenet}
\alias{bigstatsenet}
\title{Elastic-net for compressed sensing using bigstatsr}
\usage{
bigstatsenet(need, train, seed = 1218, alpha)
}
\arguments{
\item{need}{vector, numeric vector of gene that is needed}

\item{train}{matrix, numeric expression matrix of training data}

\item{seed}{numeric, random seed}

\item{alpha}{numeric, between 0 and 1}
}
\value{
list with coefficients and predictive R2
}
\description{
The function runs elastic net to train compression for a single gene
}
