% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateError.R
\name{calculateError}
\alias{calculateError}
\title{Find errors in estimated proportion matrix and signature}
\usage{
calculateError(estProp, estSig, trueProp = NULL, yref)
}
\arguments{
\item{estProp}{matrix, numeric expression matrix of compartment proportions}

\item{estSig}{matrix, numeric matrix of compartment gene signatures}

\item{trueProp}{matrix, known proportion matrix, NULL if not known}

\item{yref}{matrix, numeric matrix of original expression matrix}
}
\value{
reconstruction MSE
}
\description{
The function finds the error in a DeCompress-esimated proportion
and signature matrix using either known cell-type proportions or
the total gene expression matrix
}
