% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csDeCompress.R
\name{csDeCompress}
\alias{csDeCompress}
\title{TOAST-NMF deconvolution}
\usage{
csDeCompress(
  Y_raw,
  K,
  nMarker = 1000,
  FUN = nmfOut,
  InitMarker = NULL,
  TotalIter = 30,
  bound_negative = FALSE
)
}
\arguments{
\item{Y_raw}{matrix, numeric expression matrix}

\item{K}{integer, number of cell-types}

\item{nMarker}{integer, max number of genes for deconv, default 1000}

\item{FUN}{function, defaults to nmfOut}

\item{InitMarker}{vector, initial marker vector smaller than nrow of Y_raw}

\item{bound_negative}{logical, T/F to bound negative params at 0}
}
\value{
list with cell-type proportions and expression profiles
}
\description{
The function runs TOAST iteratively with NMF to find
compartment proportions and signatures
}
