% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateUnmix.R
\name{estimateUnmix}
\alias{estimateUnmix}
\title{Run unmix from DESeq2}
\usage{
estimateUnmix(yref, sigs, shift.range = 10^(1:10), logTransform = F)
}
\arguments{
\item{yref}{matrix, numeric expression matrix}

\item{sigs}{matrix, numeric expression compartment-specific sig matrix}

\item{shift.range}{vector, numeric vector for gene expression shifts}

\item{logTransform}{logical, T/F to logTransform data, defaults to F}
}
\value{
list with cell-type proportions and expression profiles
}
\description{
The function estimates the compartment proportions from an estimated
signature matrix using unmix
}
