% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandTarget.R
\name{expandTarget}
\alias{expandTarget}
\title{Expand targetted panel to larger feature space}
\usage{
expandTarget(target, compression_mat)
}
\arguments{
\item{target}{matrix, numeric expression matrix from targetted panel}

\item{compression_mat}{matrix, compression matrix}
}
\value{
expanded expression matrix
}
\description{
The function multiplies the targetted panel with the compression
matrix for the larger feature space
}
