% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInformSet.R
\name{findInformSet}
\alias{findInformSet}
\title{Select the compartment specific genes}
\usage{
findInformSet(
  yref,
  method = c("variance", "linearity"),
  n_genes = 1000,
  n.types = NULL,
  scree = "cumvar"
)
}
\arguments{
\item{yref}{matrix, numeric expression matrix}

\item{method}{character, variance/linearity for TOAST or LINSEED}

\item{n_genes}{integer, number of CTS genes needed}

\item{n.types}{integer, number of compartments}

\item{scree}{character, method to estimate n.types}
}
\value{
numeric matrix of n_genes CTS genes
}
\description{
The function feature selects the compartment specific genes
from the reference expression matrix
}
