% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterateSigs.R
\name{iterateSigs}
\alias{iterateSigs}
\title{Use NMF repeatedly with TOAST framework to find cell-type gene signatures
and do deconvolution}
\usage{
iterateSigs(
  yref,
  n.types = NULL,
  n.markers = 1000,
  n.iter = 30,
  scree = c("drop", "cumvar", "residual"),
  logTransform = F
)
}
\arguments{
\item{yref}{matrix, numeric expression matrix}

\item{n.types}{integer, number of cell-types}

\item{scree}{character, method to estimate n.types if n.types is NULL}

\item{iters}{numeric, number of interations}

\item{pval}{numeric, p-value cutoff}

\item{log}{logical, T/F if yref is in log-scale}
}
\value{
list with cell-type proportions and expression profiles
}
\description{
The function finds cell-type gene signatures and does deconvolution
using the TOAST framework with NMF
}
