% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaiser.R
\name{kaiser}
\alias{kaiser}
\title{Kaiser method to determine number of important SVs}
\usage{
kaiser(varpc, low = 0.08, max.pc = 0.9)
}
\arguments{
\item{varpc}{vector, singular values of a matrix}

\item{low}{numeric, lower bound of cumulative var explained}

\item{max.pc}{numeric, upper bound of cumulative var explained}
}
\value{
number of cell types
}
\description{
The function estimates the SV for the elbow in a scree plot
from the vector of singular values
}
