% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2Magic.R
\name{l2Magic}
\alias{l2Magic}
\title{Elastic-net for compressed sensing}
\usage{
l2Magic(need, train, seed = 1218, lambda = 0.1)
}
\arguments{
\item{need}{vector, numeric vector of gene that is needed}

\item{train}{matrix, numeric expression matrix of training data}

\item{seed}{numeric, random seed}

\item{lambda}{numeric, penalty parameter}
}
\value{
list with coefficients and predictive R2
}
\description{
The function runs nonlinear optimization of l2-penalty
to train compression for a single gene
}
