% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linCor.R
\name{linCor}
\alias{linCor}
\title{Use mutual linearity to find cell-type gene signatures and do deconvolution}
\usage{
linCor(
  yref,
  iters = 100,
  pval = 0.01,
  n.types = NULL,
  scree = "cumvar",
  logTransform = F
)
}
\arguments{
\item{yref}{matrix, numeric expression matrix}

\item{iters}{numeric, number of interations}

\item{pval}{numeric, p-value cutoff}

\item{n.types}{integer, number of cell-types}

\item{scree}{character, method to estimate n.types if n.types is NULL}

\item{logTransform}{logical, T/F if yref is in log-scale}
}
\value{
list with cell-type proportions and expression
}
\description{
The function finds cell-type gene signatures and does deconvolution
using the linseed pacakge
}
