% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseError.R
\name{mseError}
\alias{mseError}
\title{Select the number of cell types using SVD methods}
\usage{
mseError(mat1, mat2, rand = F, r2 = F)
}
\arguments{
\item{mat1}{matrix, numeric matrix of proportions}

\item{mat2}{matrix, numeric matrix of proportions}

\item{rand}{logical, T/F to shuffle columns of mat2}
}
\value{
number of cell types
}
\description{
The function estimates the number of cell-types using one
of three SVD methods on the input expression matrix
}
