% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfOut.R
\name{nmfOut}
\alias{nmfOut}
\title{NMF function for TOAST}
\usage{
nmfOut(Y, K)
}
\arguments{
\item{Y}{matrix, numeric expression matrix}

\item{K}{integer, number of cell-types or rank of factorization}
}
\value{
coefficient matrix from NMF
}
\description{
The function runs NMF on an expression matrix with rank K
}
