% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainCS.R
\name{trainCS}
\alias{trainCS}
\title{Wrapper to train compression matrix}
\usage{
trainCS(
  yref,
  yref_need,
  seed = 1218,
  method = c("lar", "lasso", "enet", "ridge", "l1", "TV", "l2"),
  par = T,
  n.cores,
  lambda = 0.1
)
}
\arguments{
\item{yref}{matrix, numeric expression matrix for target genes}

\item{yref_need}{matrix, numeric expression matrix of needed genes}

\item{seed}{numeric, random seed}

\item{method}{vector, character vector of optimization methods}

\item{par}{logical, T/F for parallelization}

\item{n.cores}{numeric, number of cores}

\item{lambda}{numeric, penalty paramemter for non-linear optimization}
}
\value{
list with coefficients and predictive R2
}
\description{
The function runs trainCS_gene over the entire set of genes needed
for deconvolution from the reference dataset.
}
