% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainCS_gene.R
\name{trainCS_gene}
\alias{trainCS_gene}
\title{Wrapper to train compression matrix}
\usage{
trainCS_gene(
  need,
  train,
  seed,
  method = c("lar", "lasso", "enet", "ridge", "l1", "TV", "l2"),
  n.cores,
  lambda = 0.1
)
}
\arguments{
\item{need}{vector, numeric expression vector for a needed gene}

\item{train}{matrix, numeric expression matrix for target genes}

\item{seed}{numeric, random seed}

\item{method}{vector, character vector of optimization methods}

\item{n.cores}{numeric, number of cores}

\item{lambda}{numeric, penalty paramemter for non-linear optimization}

\item{par}{logical, T/F for parallelization}
}
\value{
list with coefficients and predictive R2
}
\description{
The function estimates the best set of coefficients for a given
gene from the target genes.
}
