% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vardecomp.R
\name{vardecomp}
\alias{vardecomp}
\title{Select CTS genes with variance proporties}
\usage{
vardecomp(yref, n_genes = 1000, n.types)
}
\arguments{
\item{yref}{matrix, numeric matrix of reference expression}

\item{n_genes}{integer, number of genes requested}
}
\value{
numeric matrix of n_genes CTS genes
}
\description{
The function runs TOAST's findRefinx function to find
a set of top cell-type specific genes
}
